+++ bankATM/src/bankATM/InterestController.java
@@ -2,8 +2,8 @@ package bankATM;
 
 public interface InterestController {
 	
-	public abstract void setSavingsInterest(double interest);
+	public abstract void setSavingsInterest(float interest);
 	
-	public abstract void setLoanInterest(double interest);
+	public abstract void setLoanInterest(float interest);
 	
 }
